arm64: dts: qcom: sm8750: Add BWMONs (#11)
* dt-bindings: interconnect: qcom,msm8998-bwmon: Allow 'nonposted-mmio'

One of the BWMON instances on SM8750 requires that its MMIO space is
mapped specifically with the nE memory attribute.

Allow the nonposted-mmio property which instructs the OS to do so.

Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>

* arm64: dts: qcom: sm8750: Add BWMONs

Add the CPU BWMONs for SM8750 SoCs.

Notably, the one related to cluster0 requires that it's mapped with
the nE memory attribute. This is specific to a single instance, on this
platform only and should not be mimicked elsewhere.

Signed-off-by: Shivnandan Kumar <quic_kshivnan@quicinc.com>
[konrad: add nonposted-mmio where necessary, re-sort nodes]
Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>

* Adding ak09970 for tri key on OnePlus13 (3 step slider on left of device)

---------

Signed-off-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>
Signed-off-by: Shivnandan Kumar <quic_kshivnan@quicinc.com>
Co-authored-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>
Co-authored-by: Shivnandan Kumar <quic_kshivnan@quicinc.com>
Co-authored-by: root <root@localhost.localdomain>